requirements.txt
Pillow
numpy
scikit-learn
tqdm